gestion de la page d'accueil
