+++ pages/index.tsx
@@ -1,383 +1,57 @@
-import React from "react";
+import { useRouter } from "next/router";
 import { GetServerSideProps } from "next";
-import { categoryTab, sections, categoryObject } from "../src/data/mag";
-type Products = {
-  id?: any;
-  label?: string;
-  url?: string;
-  x?: any;
-  y?: any;
-  category?: any;
-  bgColor?: string;
-};
 
-type mapProps = {
-  products: Products[];
-  search: any;
-};
+const Acceuil: React.FC<{ search }> = ({ search }) => {
+  const router = useRouter();
 
-const HomePage: React.FC<mapProps> = () => {
-  const [resolvedProduct, setResolvedProduct] = React.useState<Products[]>([]);
-
-  const [data, setData] = React.useState(categoryObject);
-
-  const [searchedProduct, setSearchedProduct] = React.useState("");
-
-  const [changeColor, setChangeColor] = React.useState("red");
-
-  const [idProduct, setIdProduct] = React.useState("");
-
-  const userAction = async (input: any) => {
-    await fetch(`/api/search?q=${input}`, {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-    })
-      .then((responseApi) => responseApi.json())
-      .then((result) => {
-        result.positionTab.map(({ x, y, label }) => {
-          sections.map((category) => {
-            if (
-              category.left <= x &&
-              x <= category.width &&
-              category.top <= y &&
-              y <= category.height
-            ) {
-              setData({
-                ...categoryObject,
-                [category.name]: {
-                  ...categoryObject[category.name],
-                  products: result,
-                },
-              });
-            }
-          });
-        });
-      });
-  };
-
-  const dotStyle = (x: number, y: number) => {
-    return {
-      top: y,
-      left: x,
-      width: 30,
-      height: 30,
-      color: "red",
-      "background-color": "red",
-      "border-radius": 25,
-      animation: "Test 1s infinite",
-    };
-  };
-  const dotStyleOneProduct = (x: number, y: number) => {
-    return {
-      top: y,
-      left: x,
-      width: 30,
-      height: 30,
-      color: "blue",
-      "background-color": "blue",
-      "border-radius": 25,
-      animation: "Test 1s infinite",
-    };
-  };
-  const imageScale = () => {
-    return {
-      transform: "scale(1.0056646525679758)",
-    };
-  };
   return (
     <>
-      <nav className="notreNav navbar navbar-expand-lg navbar-light bg-light">
-        <div className="container-fluid">
-          <a className="navbar-brand" href="#">
-            Product Locator
-          </a>
-          <button
-            className="navbar-toggler"
-            type="button"
-            data-bs-toggle="collapse"
-            data-bs-target="#navbarScroll"
-            aria-controls="navbarScroll"
-            aria-expanded="false"
-            aria-label="Toggle navigation"
-          >
-            <span className="navbar-toggler-icon"></span>
-          </button>
-          <div className="collapse navbar-collapse" id="navbarScroll">
-            <ul className="navbar-nav me-auto my-2 my-lg-0 navbar-nav-scroll">
-              <li className="nav-item">
-                <a className="nav-link active" aria-current="page" href="#">
-                  Home
-                </a>
-              </li>
-            </ul>
-            <a className="btn btn-outline-success me-2" type="button" href="#">
-              <div className="bd-highlight">
-                <i className="fas fa-store-alt"></i>
-              </div>
-            </a>
-            <a
-              className="btn btn-outline-success me-2"
-              type="button"
-              data-bs-toggle="modal"
-              data-bs-target="#exampleModal"
-              data-bs-whatever="@getbootstrap"
-            >
-              <div className=" bd-highlight">
-                <i className="far fa-user"></i>
-              </div>
-            </a>
-
-            <div className="d-flex">
-              <input
-                className="form-control me-2"
-                type="search"
-                placeholder="Search"
-                name="searchInput"
-                aria-label="Search"
-                onChange={(e) => setSearchedProduct(e.target.value)}
-              />
-              <button
-                className="btn btn-outline-success"
-                type="submit"
-                onClick={() => userAction(searchedProduct)}
-              >
-                Search
-              </button>
-            </div>
-          </div>
-        </div>
-      </nav>
-      {/* Login page */}
       <div
-        className="modal fade"
-        id="exampleModal"
-        aria-labelledby="exampleModalLabel"
-        aria-hidden="true"
+        className="d-flex flex-column justify-content-center align-items-center align-middle"
+        style={{ margin: "13rem 0 0 0" }}
       >
-        <div className="modal-dialog">
-          <div className="modal-content">
-            <div className="modal-header">
-              <button
-                type="button"
-                className="btn-close"
-                data-bs-dismiss="modal"
-                aria-label="Close"
-              ></button>
-            </div>
-            <div className="modal-body">
-              <div className="d-flex justify-content-center h-100">
-                <div className="card">
-                  <div className="card-header">
-                    <h3>Sign In</h3>
-                    <div className="d-flex justify-content-end social_icon">
-                      <span>
-                        <i className="fab fa-facebook-square"></i>
-                      </span>
-                      <span>
-                        <i className="fab fa-google-plus-square"></i>
-                      </span>
-                      <span>
-                        <i className="fab fa-twitter-square"></i>
-                      </span>
-                    </div>
-                  </div>
-                  <div className="card-body">
-                    <form>
-                      <div className="row mb-3">
-                        <label className="col-sm-2 col-form-label">Email</label>
-                        <div className="col-sm-10">
-                          <input
-                            type="email"
-                            className="form-control"
-                            id="inputEmail3"
-                          />
-                        </div>
-                      </div>
-                      <div className="row mb-3">
-                        <label className="col-sm-2 col-form-label">
-                          Password
-                        </label>
-                        <div className="col-sm-10">
-                          <input
-                            type="password"
-                            className="form-control"
-                            id="inputPassword3"
-                          />
-                        </div>
-                      </div>
-                    </form>
-                  </div>
-                  <div className="card-footer">
-                    <div className="d-flex justify-content-center links">
-                      Don't have an account?<a href="#">Sign Up</a>
-                    </div>
-                    <div className="d-flex justify-content-center">
-                      <a href="#">Forgot your password?</a>
-                    </div>
-                  </div>
-                </div>
-              </div>
-            </div>
-            <div className="modal-footer">
-              <button
-                type="button"
-                className="btn btn-secondary"
-                data-bs-dismiss="modal"
+        <div className="">
+          <img
+            src="/accueil.png"
+            className="rounded "
+            alt="..."
+            style={{ width: "30rem" }}
+          />
+        </div>
+
+        <div className="input-group w-50">
+          <form action="/api/accueil" method="POST">
+            <input
+              type="text"
+              name="input"
+              className="form-control p-3 mt-3"
+              placeholder="Search product"
+              aria-label="Input group example"
+              aria-describedby="basic-addon1"
+            />
+            <button
+              className="input-group-text p-3 mt-3 "
+              style={{ cursor: "pointer" }}
+              type="submit"
+              onSubmit={(e) => e.preventDefault}
+            >
+              <svg
+                xmlns="http://www.w3.org/2000/svg"
+                width="16"
+                height="16"
+                fill="currentColor"
+                className="bi bi-search "
+                viewBox="0 0 16 16"
               >
-                Close
-              </button>
-            </div>
-          </div>
+                <path d="M11.742 10.344a6.5 6.5 0 1 0-1.397 1.398h-.001c.03.04.062.078.098.115l3.85 3.85a1 1 0 0 0 1.415-1.414l-3.85-3.85a1.007 1.007 0 0 0-.115-.1zM12 6.5a5.5 5.5 0 1 1-11 0 5.5 5.5 0 0 1 11 0z"></path>
+              </svg>
+            </button>
+          </form>
         </div>
-      </div>
-
-      <button onClick={() => console.log(window.innerWidth / 2648)}></button>
-
-      {/* Fin login page */}
-
-      <div className="image" style={{ transform: `scale(0.5438066465256798)` }}>
-        <img
-          src="https://m1.lmcdn.fr/media/18/5d0901067eb45f348d8a8c9f/1798045962/map-png-store-3.png"
-          alt=""
-        />
-        {sections.map(({ name }) => {
-          return data[name].products?.positionTab?.map((product, index) => {
-            return (
-              <>
-                {product.id === idProduct ? (
-                  <div
-                    key={index}
-                    className={name}
-                    style={dotStyleOneProduct(product.x, product.y)}
-                  ></div>
-                ) : (
-                  <div
-                    key={index}
-                    className={name}
-                    style={dotStyle(product.x, product.y)}
-                  ></div>
-                )}
-              </>
-            );
-          });
-        })}
-
-        <div className="liste">
-          <button
-            className="btn btn-primary"
-            type="button"
-            data-bs-toggle="offcanvas"
-            data-bs-target="#offcanvasExample"
-            aria-controls="offcanvasExample"
-          >
-            <h4>PRODUCTS</h4>
-          </button>
-          <div
-            className="offcanvas offcanvas-start"
-            id="offcanvasExample"
-            aria-labelledby="offcanvasExampleLabel"
-            style={{ height: 1950, width: 600 }}
-          >
-            <div className="offcanvas-header">
-              <h3 className="offcanvas-title" id="offcanvasExampleLabel">
-                Product Locator
-              </h3>
-              <button
-                type="button"
-                className="btn-close text-reset"
-                data-bs-dismiss="offcanvas"
-                aria-label="Close"
-              ></button>
-            </div>
-            <div className="offcanvas-body">
-              <div className="col-xl-auto">
-                <div className="col-xl-auto" style={{ width: 16, height: 27 }}>
-                  {sections.map(({ name }) => {
-                    return data[name].products?.positionTab?.map(
-                      (product, index) => {
-                        return (
-                          <>
-                            <div
-                              className="card d-flex justify-content-evenly"
-                              style={{ width: 500, height: 270 }}
-                              key={index}
-                            >
-                              <div className="card-body">
-                                <br />
-                                <h4 className="card-title">{product.label}</h4>
-                                <br />
-                                <img
-                                  src={product.img}
-                                  className="card-img-top"
-                                  alt="my Image"
-                                  style={{ width: 120, height: 100 }}
-                                />
-                                <br />
-                                <h5 className="card-title">
-                                  {product.price + "0"} €
-                                </h5>
-                                <button
-                                  className="btn btn-primary"
-                                  onClick={(e) => setIdProduct(product.id)}
-                                  value={product.id}
-                                >
-                                  Afficher ce produit sur la carte
-                                </button>
-
-                                <br />
-                              </div>
-                              <br />
-                            </div>
-                          </>
-                        );
-                      }
-                    );
-                  })}
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
+        <h1 className="mt-5">Chaque recherche vous fait gagner du temps</h1>
       </div>
     </>
   );
 };
 
-export default HomePage;
-
-export const getServerSideProps: GetServerSideProps = async () => {
-  /////////////////////////////////////////////////////////////////////
-  //                            GET ALL STORES                       //
-  ////////////////////////////////////////////////////////////////////
-  await fetch(`https://api-gateway.leroymerlin.fr/api-stock/v1/stores/all`, {
-    method: "GET",
-    headers: {
-      "X-Gateway-APIKey": `${process.env.TOKEN_API_STOCK}`,
-    },
-  })
-    .then((response) => response.json())
-    .then((stores) => {
-      //console.log(stores);
-    });
-
-  ///////////////////////////////////////////////////////////////////////
-  //        GET PRODUCTS LOCALISATION IN STORE                      //
-  ///////////////////////////////////////////////////////////////////
-
-  await fetch(` https://api-gateway.leroymerlin.fr/api-geoproduct/v2/stores/`, {
-    method: "GET",
-    headers: {
-      "X-Gateway-APIKey": `${process.env.TOKEN_API_GEOPRODUCT}`,
-    },
-  })
-    .then((response) => response.json())
-    .then((stores) => {
-      // console.log(stores);
-    });
-
-  return {
-    props: {},
-  };
-};
+export default Acceuil;
